requirements.txt
streamlit
google-generativeai
numpy
pypdf
python-dotenv
chromadb
sentence-transformers
langchain
